Magnum: Added recording of shots to every app
