+++ src/robot_dart/gui/magnum/glx_application.hpp
@@ -3,9 +3,7 @@
 
 #include <robot_dart/gui/magnum/base_application.hpp>
 
-#include <Corrade/Containers/Optional.h>
 #include <Magnum/GL/Framebuffer.h>
-#include <Magnum/Image.h>
 
 namespace robot_dart {
     namespace gui {
@@ -17,16 +15,11 @@ namespace robot_dart {
 
                 void render() override;
 
-                Corrade::Containers::Optional<Magnum::Image2D>& image() { return _image; }
-
             protected:
                 Magnum::GL::Framebuffer _framebuffer{Magnum::NoCreate};
-                Corrade::Containers::Optional<Magnum::Image2D> _image;
                 // size_t _index = 0;
 
                 virtual int exec() override { return 0; }
-
-                Corrade::Containers::Optional<Magnum::PixelFormat> _getPixelFormat(Magnum::GL::AbstractFramebuffer& framebuffer);
             };
         } // namespace magnum
     } // namespace gui
